Swap over to the new stuff
