+++ nand/test_component.py
@@ -1,79 +1,116 @@
-from random import shuffle
-
-from nand.component import sorted_nodes
+from nand.component import *
 
 
-class Node:
-    """Trivial node type providing refs() and inst, as required by sorted_nodes."""
-    
-    def __init__(self, name, *rs):
-        self.name = name
-        self.rs = set(rs)
-        self.inst = self
-    
-    def refs(self):
-        lst = list(self.rs)
-        shuffle(lst)
-        return lst
-        
-    def __repr__(self):
-        return self.name
-        
-def test_simple():
-    a = Node('a')
-    b = Node('b', a)
-    c = Node('c', b)
-    
-    assert sorted_nodes(c) == [a, b, c]
-    
-    
-def test_converging():
-    a = Node('a', )
-    b = Node('b', a)
-    c = Node('c', a, b)
-    
-    assert sorted_nodes(c) == [a, b, c], "a should precede b, even if c also depends directly on a"
+def test_nand():
+    nand = Nand()
 
-    
-def test_circular():
-    a = Node('a')
-    b = Node('b', a)
-    c = Node('c', b)
-    a.rs.add(c)
-    
-    assert sorted_nodes(c) == [a, b, c], "a is first because we started at c"
+    assert nand.inputs() == {"a": 1, "b": 1}
+    assert nand.outputs() == {"out": 1}
+
+    # assert nand.wire({a: 0b001, b: 0b010}) == {}
+
+    traces = {"a": [0b001], "b": [0b010], "out": [0b100]}
+
+    op, = nand.combine(**traces)
+    assert op(0b000) == 0b100
+    assert op(0b001) == 0b101
+    assert op(0b010) == 0b110
+    assert op(0b011) == 0b011
+    assert op(0b100) == 0b100
+    assert op(0b101) == 0b101
+    assert op(0b110) == 0b110
+    assert op(0b111) == 0b011
+
+    assert nand.sequence(**traces) == []
 
 
-# This isn't quite right, because:
-# 1) The dependencies on Mux_0 and Mux_8 are missing 
-# 2) Would need to account for forward (back) refs in the assertion also
-# def test_complex():
-#     # based on the failing (flaky) test for Bit:
-#     # inst: Not_1; {'in_': Ref(root.load)}
-#     # inst: Not_9; {'in_': Ref(root.clock)}
-#     # inst: Mux_0; {'a': Ref(Forward(DFF_6).out), 'b': Ref(root.in_), 'sel': Ref(root.load)}
-#     # inst: Mux_8; {'a': Ref(Forward(Mux_8).out), 'b': Ref(Mux_0.out), 'sel': Ref(root.clock)}
-#     # inst: Not_18; {'in_': Ref(Not_14.out)}
-#     # inst: Not_14; {'in_': Ref(root.clock)}
-#     # inst: Latch_7; {'in_': Ref(Mux_0.out), 'enable': Ref(root.clock)}
-#     # inst: Mux_17; {'a': Ref(Forward(Mux_17).out), 'b': Ref(Latch_7.out), 'sel': Ref(Not_14.out)}
-#     # inst: Latch_16; {'in_': Ref(Latch_7.out), 'enable': Ref(Not_14.out)}
-#     # inst: DFF_6; {'in_': Ref(Mux_0.out), 'clock': Ref(root.clock)}
-#     a = Node('Not_1')
-#     b = Node('Mux_0')
-#     c = Node('Not_9')
-#     d = Node('Mux_8', b); d.rs.add(d)
-#     e = Node('Latch_7', b)
-#     f = Node('Not_14')
-#     g = Node('Not_18', f)
-#     h = Node('Mux_17', e); h.rs.add(h)
-#     i = Node('Latch_16', e, f)
-#     j = Node('DFF_6', b)
-#
-#     # delayed:
-#     b.rs.add(j)
-#
-#     s = sorted_nodes(b)
-#     for n in (a, b, c, d, e, f, g, h, i, j):
-#         for r in n.refs():
-#             assert s.index(r) <= s.index(n)
+def test_dff():
+    dff = DFF()
+
+    assert dff.inputs() == {"in_": 1}
+    assert dff.outputs() == {"out": 1}
+
+    traces = {"in_": [0b01], "out": [0b10]}
+
+    assert dff.combine(**traces) == []
+
+    op, = dff.sequence(**traces)
+    assert op(0b00) == 0b00
+    assert op(0b01) == 0b11
+    assert op(0b10) == 0b00
+    assert op(0b11) == 0b11
+
+
+def test_rom():
+    rom = ROM(4)
+
+    assert rom.inputs() == {"address": 4}
+    assert rom.outputs() == {"out": 16}
+
+    traces = {
+        "address": [0b0001 << i for i in range(4)],
+        "out": [0b10000 << i for i in range(16)]
+    }
+
+    op, = rom.combine(**traces)
+
+    rom.program([1, 2, 3, 4, 5])
+    for i in range(16):
+        if i < 5:
+            assert op(i) == ((i+1) << 4) | i
+        else:
+            assert op(i) == 0x00 | i
+
+    assert rom.sequence(**traces) == []
+
+
+def test_ram():
+    ram = RAM(4)
+
+    assert ram.inputs() == {"in_": 16, "load": 1, "address": 4}
+    assert ram.outputs() == {"out": 16}
+
+    traces = {
+        "in_": [0b1 << i for i in range(16)],
+        "load": [0b1 << 16],
+        "address": [0b1 << (i+17) for i in range(4)],
+        "out": [0b1 << (i+21) for i in range(16)]
+    }
+
+    op, = ram.combine(**traces)
+
+    for i in range(16):
+        ram.set(i, i)
+
+    for i in range(16):
+        addr = i << 17
+        out = i << 21
+        assert op(addr) == out | addr
+
+
+    op, = ram.sequence(**traces)
+
+    for i in range(16):
+        in_ = 12345 + i
+        load = 0b1 << 16
+        addr = i << 17
+        op(addr | load | in_)
+        assert ram.get(i) == 12345 + i
+
+
+def test_input():
+    inpt = Input()
+
+    assert inpt.inputs() == {}
+    assert inpt.outputs() == {"out": 16}
+
+    traces = { "out": [0b1 << i for i in range(16)]}
+
+    op, = inpt.combine(**traces)
+
+    assert op(0) == 0
+
+    inpt.set(12345)
+    assert op(0) == 12345
+
+    assert inpt.sequence(**traces) == []
