--- nand_new/__init__.py
@@ -1,13 +1,0 @@
-"""DSL for defining components in the style of Nand to Tetris.
-
-See project_01.py and test_01.py for examples of how to use it.
-
-This module just re-exports the definitions that are most often needed from where they're actually
-implemented (in the `eval` sub-dir.)
-"""
-
-# from nand.component import Component, Const, Nand, DynamicDFF, Memory, clock, lazy, gate_count, unsigned
-# from nand.compiler import run
-
-from nand.component import unsigned
-from nand_new.syntax import Nand, build, run, gate_count
