+++ project_02.py
@@ -1,5 +1,6 @@
-from nand_new import *
+from nand import *
 from project_01 import And, And16, Or, Mux16, Not, Not16, Xor
+
 
 def mkHalfAdder(inputs, outputs):
     a = inputs.a
